[Android] Use ant and xwalk_core_library for make_apk
Previously, we are copying the python script from gyp
and reproduce the steps gyp generated to build
xwalk_app_template_apk to make apk.

It's hard to maintain and now xwalk_core_library is
mature enough to be used as Android Library Project.

This commit uses ant and xwalk_core_library to make
apk to replace previous approach.

The changes including:
1. The app packaging tool structure:
 a. Assets, jars and template code are now put in
    "template", jars includes xwalk_runtime_java
    and xwalk_app_runtime_java.
 b. xwalk_core_library will be put in root folder of
    packaging tool.
 c. Scripts used by gyp generated commands are removed.
2. When make apk:
 a. Embedded Mode:
  1) Copy template to app folder named according to
     input app name.
  2) Copy xwalk_core_library into app folder.
  3) Move arch folders under "app/xwalk_core_library/libs/"
     to "app/native_libs/<arch>"
  4) For each target arch:
   4.1) Run "android update lib-project" for app/xwalk_core_library.
   4.2) Run "android update project" for app.
   4.3) Remove all subfolder in "app/xwalk_core_library/libs",
        to make sure it will only contains one arch folder after 4.4).
   4.4) Move back the needed native library for target arch back.
        it's actually partly reverting step 3).
   4.5) Run "ant release" in app.
   4.6) Copy and rename the built apk out to root folder of
        packaging tool.
 b. Shared Mode:
   1) Copy template to app folder named according to
      input app name.
   2) Remove xwalk_runtime_java.jar in "app/libs".
   3) Run "android update project" for app
   5) Run "ant release" in app.
   6) Copy and rename the built apk out to root folder of packaging
      tool.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2128

+++ app/tools/android/customize.py
@@ -116,9 +116,9 @@ def Prepare(app_info, compressor):
   app_root = app_info.app_root
   if os.path.exists(name):
     shutil.rmtree(name)
-  shutil.copytree('app_src', name)
+  shutil.copytree('template', name)
   shutil.rmtree(os.path.join(name, 'src'))
-  src_root = os.path.join('app_src', 'src', 'org', 'xwalk', 'app', 'template')
+  src_root = os.path.join('template', 'src', 'org', 'xwalk', 'app', 'template')
   src_activity = os.path.join(src_root, 'AppTemplateActivity.java')
   if not os.path.isfile(src_activity):
     print ('Please make sure that the java file'
@@ -130,20 +130,19 @@ def Prepare(app_info, compressor):
   dest_activity = name + 'Activity.java'
   shutil.copyfile(src_activity, os.path.join(root_path, dest_activity))
   if app_root:
-    assets_path = os.path.join(name, 'assets')
-    shutil.rmtree(assets_path)
-    os.makedirs(assets_path)
-    app_src_path = os.path.join(assets_path, 'www')
-    shutil.copytree(app_root, app_src_path)
+    assets_path = os.path.join(name, 'assets', 'www')
+    if os.path.isdir(assets_path):
+      shutil.rmtree(assets_path)
+    shutil.copytree(app_root, assets_path)
     if compressor:
-      CompressSourceFiles(app_src_path, compressor)
+      CompressSourceFiles(assets_path, compressor)
 
 
 def CustomizeStringXML(name, description):
   strings_path = os.path.join(name, 'res', 'values', 'strings.xml')
   if not os.path.isfile(strings_path):
     print ('Please make sure strings_xml'
-           ' exists under app_src folder.')
+           ' exists under template folder.')
     sys.exit(6)
 
   if description:
@@ -185,7 +184,7 @@ def CustomizeXML(app_info, description, icon_dict, manifest, permissions):
   manifest_path = os.path.join(name, 'AndroidManifest.xml')
   if not os.path.isfile(manifest_path):
     print ('Please make sure AndroidManifest.xml'
-           ' exists under app_src folder.')
+           ' exists under template folder.')
     sys.exit(6)
 
   CustomizeStringXML(name, description)
